fix(mobile): split widget plugin helper
- Move AppDelegate patcher

- Satisfy file-size gate
